# demo/schema.openapi.yaml
openapi: 3.0.0
servers:
  - url: http://xkcd.com/
info:
  description: Webcomic of romance, sarcasm, math, and language.
  title: XKCD
  version: 1.0.0
  x-apisguru-categories:
    - media
  x-logo:
    url: http://imgs.xkcd.com/static/terrible_small_logo.png
  x-origin:
    - format: openapi
      url: https://raw.githubusercontent.com/APIs-guru/unofficial_openapi_specs/master/xkcd.com/1.0.0/openapi.yaml
      version: "3.0"
  x-providerName: xkcd.com
  x-tags:
    - humor
    - comics
  x-unofficialSpec: true
externalDocs:
  url: https://xkcd.com/json.html
paths:
  /info:
    get:
      description: |
        Fetch current comic and metadata.
      security:
        - authHeader: []
      responses:
        "200":
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/comic"
          description: OK
        default:
          description: Any error
          content:
            "application/problem+json":
              schema:
                $ref: "#/components/schemas/problem"
  "/{comicId}/info":
    get:
      description: |
        Fetch comics and metadata  by comic id.
      security:
        - authHeader: []
      parameters:
        - in: path
          name: comicId
          required: true
          schema:
            type: integer
            minimum: 1
            exclusiveMinimum: false
            maximum: 2147483648
      responses:
        "200":
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/comic"
          description: OK
        default:
          description: Any error
          content:
            "application/problem+json":
              schema:
                $ref: "#/components/schemas/problem"
components:
  schemas:
    comic:
      properties:
        alt:
          type: string
        day:
          type: string
        img:
          type: string
        link:
          type: string
        month:
          type: string
        news:
          type: string
        num:
          type: number
        safe_title:
          type: string
        title:
          type: string
        transcript:
          type: string
        year:
          type: string
      type: object
    problem:
      type: object
      required:
        - status
        - title
      additionalProperties: true
      properties:
        status:
          type: integer
          description: The HTTP status code generated by the origin server for this occurrence of the problem
          minimum: 100
          maximum: 599
          example: 500
        title:
          type: string
          description: A short, human-readable summary of the problem type. It should not change from occurrence to occurrence of the problem, except for purposes of localization.
          example: "Internal server error"
        type:
          type: string
          description: A URI reference that identifies the problem type.
          format: uri
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence of the problem.
        instance:
          type: string
          description: A URI reference that identifies the specific occurrence of the problem. It may or may not yield further information if dereferenced.
  securitySchemes:
    authHeader:
      type: http
      scheme: bearer